Addded: Individual Sensor PWR Control adn Reading.
Sending Read Data to MQTT (Includes TBD and GPS)

[TESTED OK]

+++ lib/sensor_ctrl/sensor_ctrl.h
@@ -6,7 +6,7 @@
 struct sensor_packet
 {
     int turbidity; //Stored in NTUs
-    int longitude, lattitude;
+    float longitude, lattitude;
 };
 
 extern struct k_msgq sensor_msgq;
